+++ packages/api/tests/endpoints/test-executions.js
@@ -1196,7 +1196,7 @@ test.serial('POST /executions/workflows-by-granules returns correct workflows wh
   t.deepEqual(response.body.sort(), ['fakeWorkflow', 'workflow2']);
 });
 
-test.serial('POST /executions creates a new execution in DynamoDB/PostgreSQL/Elasticsearch with correct timestamps', async (t) => {
+test.serial('POST /executions creates a new execution in PostgreSQL/Elasticsearch with correct timestamps', async (t) => {
   const newExecution = fakeExecutionFactoryV2();
 
   await request(app)
@@ -1205,10 +1205,6 @@ test.serial('POST /executions creates a new execution in DynamoDB/PostgreSQL/Ela
     .set('Accept', 'application/json')
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
-
-  const fetchedDynamoRecord = await executionModel.get({
-    arn: newExecution.arn,
-  });
 
   const fetchedPgRecord = await t.context.executionPgModel.get(
     t.context.knex,
@@ -1219,16 +1215,13 @@ test.serial('POST /executions creates a new execution in DynamoDB/PostgreSQL/Ela
 
   const fetchedEsRecord = await t.context.esExecutionsClient.get(newExecution.arn);
 
-  t.true(fetchedDynamoRecord.createdAt > newExecution.createdAt);
-  t.true(fetchedDynamoRecord.updatedAt > newExecution.updatedAt);
+  t.true(fetchedPgRecord.created_at.getTime() > newExecution.createdAt);
 
-  t.is(fetchedPgRecord.created_at.getTime(), fetchedDynamoRecord.createdAt);
-  t.is(fetchedPgRecord.updated_at.getTime(), fetchedDynamoRecord.updatedAt);
   t.is(fetchedPgRecord.created_at.getTime(), fetchedEsRecord.createdAt);
   t.is(fetchedPgRecord.updated_at.getTime(), fetchedEsRecord.updatedAt);
 });
 
-test.serial('POST /executions creates the expected record in DynamoDB/PostgreSQL/Elasticsearch', async (t) => {
+test.serial('POST /executions creates the expected record in PostgreSQL/Elasticsearch', async (t) => {
   const newExecution = fakeExecutionFactoryV2({
     asyncOperationId: t.context.testAsyncOperation.id,
     collectionId: t.context.collectionId,
@@ -1247,10 +1240,6 @@ test.serial('POST /executions creates the expected record in DynamoDB/PostgreSQL
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  const fetchedDynamoRecord = await executionModel.get({
-    arn: newExecution.arn,
-  });
-
   const fetchedPgRecord = await t.context.executionPgModel.get(
     t.context.knex,
     {
@@ -1260,20 +1249,12 @@ test.serial('POST /executions creates the expected record in DynamoDB/PostgreSQL
 
   const fetchedEsRecord = await t.context.esExecutionsClient.get(newExecution.arn);
 
-  t.is(fetchedPgRecord.arn, fetchedDynamoRecord.arn);
+  t.is(fetchedPgRecord.arn, newExecution.arn);
   t.truthy(fetchedPgRecord.cumulus_id);
   t.is(fetchedPgRecord.async_operation_cumulus_id, t.context.asyncOperationCumulusId);
   t.is(fetchedPgRecord.collection_cumulus_id, t.context.collectionCumulusId);
   t.is(fetchedPgRecord.parent_cumulus_id, t.context.fakePGExecutions[1].cumulus_id);
 
-  t.deepEqual(
-    fetchedDynamoRecord,
-    {
-      ...newExecution,
-      createdAt: fetchedDynamoRecord.createdAt,
-      updatedAt: fetchedDynamoRecord.updatedAt,
-    }
-  );
   t.deepEqual(
     fetchedEsRecord,
     {
@@ -1377,10 +1358,6 @@ test.serial('POST /executions with non-existing parentArn still creates a new ex
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  const fetchedDynamoRecord = await executionModel.get({
-    arn: newExecution.arn,
-  });
-
   const fetchedPgRecord = await t.context.executionPgModel.get(
     t.context.knex,
     {
@@ -1388,7 +1365,7 @@ test.serial('POST /executions with non-existing parentArn still creates a new ex
     }
   );
 
-  t.is(fetchedPgRecord.arn, fetchedDynamoRecord.arn);
+  t.truthy(fetchedPgRecord);
   t.falsy(fetchedPgRecord.parent_cumulus_id);
 });
 
@@ -1452,7 +1429,7 @@ test.serial('POST /executions publishes message to SNS topic', async (t) => {
   t.deepEqual(executionRecord, translatedExecution);
 });
 
-test.serial('PUT /executions updates the record as expected in DynamoDB/PostgreSQL/Elasticsearch', async (t) => {
+test.serial('PUT /executions updates the record as expected in PostgreSQL/Elasticsearch', async (t) => {
   const execution = fakeExecutionFactoryV2({
     collectionId: t.context.collectionId,
     parentArn: t.context.fakeApiExecutions[1].arn,
@@ -1475,10 +1452,6 @@ test.serial('PUT /executions updates the record as expected in DynamoDB/PostgreS
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  const dynamoRecord = await executionModel.get({
-    arn: execution.arn,
-  });
-
   const pgRecord = await t.context.executionPgModel.get(
     t.context.knex,
     {
@@ -1493,58 +1466,43 @@ test.serial('PUT /executions updates the record as expected in DynamoDB/PostgreS
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  const updatedDynamoRecord = await executionModel.get({
-    arn: execution.arn,
-  });
-
-  const updatePgRecord = await t.context.executionPgModel.get(
+  const updatedPgRecord = await t.context.executionPgModel.get(
     t.context.knex,
     {
       arn: execution.arn,
     }
   );
   const updatedEsRecord = await t.context.esExecutionsClient.get(execution.arn);
-  const expectedApiRecord = {
+  const expectedEsRecord = {
     ...updatedExecution,
     collectionId: execution.collectionId,
-    createdAt: dynamoRecord.createdAt,
-    updatedAt: updatedDynamoRecord.updatedAt,
+    createdAt: updatedPgRecord.created_at.getTime(),
+    updatedAt: updatedPgRecord.updated_at.getTime(),
   };
 
-  t.deepEqual(
-    updatedDynamoRecord,
-    expectedApiRecord
-  );
   t.like(
     updatedEsRecord,
     {
-      ...expectedApiRecord,
+      ...expectedEsRecord,
       timestamp: updatedEsRecord.timestamp,
     }
   );
 
-  t.is(updatePgRecord.arn, execution.arn);
-  t.is(updatePgRecord.cumulus_id, pgRecord.cumulus_id);
+  t.is(updatedPgRecord.arn, execution.arn);
+  t.is(updatedPgRecord.cumulus_id, pgRecord.cumulus_id);
 
-  t.is(updatedDynamoRecord.createdAt, dynamoRecord.createdAt);
-  t.true(updatedDynamoRecord.updatedAt > dynamoRecord.createdAt);
-  t.is(updatePgRecord.created_at.getTime(), pgRecord.created_at.getTime());
-  t.true(updatePgRecord.updated_at.getTime() > pgRecord.updated_at.getTime());
+  t.is(updatedPgRecord.created_at.getTime(), pgRecord.created_at.getTime());
+  t.true(updatedPgRecord.updated_at.getTime() > pgRecord.updated_at.getTime());
 
   // collectionId was omitted from body of PUT request, so values are
   // not overridden in the database
-  t.is(updatedDynamoRecord.collectionId, execution.collectionId);
-  t.is(updatePgRecord.collection_cumulus_id, t.context.collectionCumulusId);
+  t.is(updatedPgRecord.collection_cumulus_id, t.context.collectionCumulusId);
   // updated record has added field
-  t.is(updatedDynamoRecord.asyncOperationId, updatedExecution.asyncOperationId);
-  t.is(updatePgRecord.async_operation_cumulus_id, t.context.asyncOperationCumulusId);
+  t.is(updatedPgRecord.async_operation_cumulus_id, t.context.asyncOperationCumulusId);
   // updated record has updated field
-  t.is(updatedDynamoRecord.parentArn, updatedExecution.parentArn);
-  t.is(updatePgRecord.parent_cumulus_id, t.context.fakePGExecutions[2].cumulus_id);
-  t.is(updatedDynamoRecord.status, updatedExecution.status);
-  t.is(updatePgRecord.status, updatedExecution.status);
-  t.deepEqual(updatedDynamoRecord.finalPayload, updatedExecution.finalPayload);
-  t.deepEqual(updatePgRecord.final_payload, updatedExecution.finalPayload);
+  t.is(updatedPgRecord.parent_cumulus_id, t.context.fakePGExecutions[2].cumulus_id);
+  t.is(updatedPgRecord.status, updatedExecution.status);
+  t.deepEqual(updatedPgRecord.final_payload, updatedExecution.finalPayload);
 });
 
 test.serial('PUT /executions throws error for arn mismatch between params and payload', async (t) => {
@@ -1648,10 +1606,6 @@ test.serial('PUT /executions with non-existing parentArn still updates the execu
     .set('Authorization', `Bearer ${jwtAuthToken}`)
     .expect(200);
 
-  const fetchedDynamoRecord = await executionModel.get({
-    arn: updatedExecution.arn,
-  });
-
   const fetchedPgRecord = await t.context.executionPgModel.get(
     t.context.knex,
     {
@@ -1659,7 +1613,7 @@ test.serial('PUT /executions with non-existing parentArn still updates the execu
     }
   );
 
-  t.is(fetchedPgRecord.arn, fetchedDynamoRecord.arn);
+  t.is(fetchedPgRecord.arn, updatedExecution.arn);
   t.falsy(fetchedPgRecord.parent_cumulus_id);
 });