+++ CHANGELOG.md
@@ -25,6 +25,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
     - Removes dynamo model calls from api/lib/granule-remove-from-cmr.unpublishGranule
     - Removes Post Deployment execution check from sf-event-sqs-to-db-records
     - Moves describeGranuleExecution from api granule model to api/lib/executions.js
+  - **CUMULUS-2806**
+    - Remove DynamoDB logic from executions `POST` endpoint
+    - Remove DynamoDB logic from sf-event-sqs-to-db-records lambda execution writes.
   - **CUMULUS-2809**
     - Remove DynamoDB logic from providers `PUT` endpoint
     - Updates DB models asyncOperation, provider and rule to return all fields on upsert.